backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Azure SDKs
azure-identity==1.15.0
azure-mgmt-resource==23.1.1
azure-mgmt-subscription==3.1.1
azure-mgmt-security==6.0.0
azure-mgmt-network==25.1.0
azure-mgmt-authorization==4.0.0
msgraph-sdk==1.0.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Reporting
jinja2==3.1.2
weasyprint==60.2
markdown==3.5.1

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
asyncio==3.4.3

# Logging and monitoring
structlog==23.2.0